tests/wpt/web-platform-tests/css/css-fonts/parsing/font-variant-caps-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-variant-caps with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-caps-prop">
<meta name="assert" content="font-variant-caps supports only the grammar 'normal | small-caps | all-small-caps | petite-caps | all-petite-caps | unicase | titling-caps'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-variant-caps', 'auto');
test_invalid_value('font-variant-caps', 'normal unicase');
</script>
</body>
</html>
